connman-plugin: be less chatty on DBus
Until now we had been simply refusing to maintain any state and just
resetting the proxy configuration every time we detect a change however
the author hadn't realised that GSettings->dconf communication happens via
DBus meaning that this simple approach is actually quite harmful.

This is a first pass at a patch to reduce the amount of DBus chatter this
plugin causes by maintaining a copy of the configuration and only changing
GSettings when the proxy configuration changes.

Fixes TZPC-3207