connman: fix comparison of ConnMan vs. GNOME proxy methods
This fixes a bug in comparing proxy configuration such that GSettings are
only changed when the configuration has changed. A direct comparison of
the method returned from ConnMan vs the method stored in GSettings was
failing because ConnMan and Gnome use different strings for direct access
('direct' vs 'none' respectively) resulting in triggering changes to the
GSettings even though the method hadn't truly changed.

Fixes TZPC-3207